Version update, minor tweaks

+++ src/main/java/sidben/ateliercanvas/client/gui/GuiCustomPaintingIconLoader.java
@@ -138,7 +138,7 @@ public abstract class GuiCustomPaintingIconLoader extends GuiScreen
     }
     
     protected float getFileSizeKBytes() {
-        return this.getFileSizeBytes() / 1024;
+        return (float)this.getFileSizeBytes() / 1024.0F;
     }
